project cleaned up
