+++ config/mainMenu.js
@@ -1,0 +1,8 @@
+const { Markup } = require('telegraf');
+
+const mainMenu = Markup.keyboard([
+    ['🎬 Get Recommendation', '🎭 Pick Mood'],
+    ['🔁 Restart Bot']
+]).resize();
+
+module.exports = { mainMenu };
